moved auto_auth view to students.view and fixed flag conflicts
