+++ lms/envs/common.py
@@ -38,7 +38,7 @@ COURSEWARE_ENABLED = True
 ENABLE_JASMINE = False
 
 # only relevant if MITX_FEATURES['AUTOMATIC_AUTH_FOR_LOAD_TESTING'] = True
-MAX_AUTO_AUTH_USERS = 20
+MAX_AUTO_AUTH_USERS = 200
 
 GENERATE_RANDOM_USER_CREDENTIALS = False
 PERFSTATS = False
